Fix mobile build
